+++ packages/terra-aggregator/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+4.29.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.28.0 - (October 3, 2019)
 ------------------
 ### Changed