+++ packages/terra-date-picker/CHANGELOG.md
@@ -4,6 +4,15 @@ Changelog
 Unreleased
 ----------
 
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+
 4.15.0 - (October 3, 2019)
 ------------------
 ### Changed