+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -4,6 +4,17 @@ ChangeLog
 Unreleased
 ----------
 
+4.16.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `isInvalid` prop and stylings
+* Added `isInvalidMeridiem` prop 
+* Added `required` prop 
+
 4.15.0 - (October 3, 2019)
 ------------------
 ### Changed