+++ packages/terra-hookshot/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+5.20.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 5.19.0 - (October 3, 2019)
 ------------------
 ### Changed